# No longer omit stdlib dependencies from dependency inference.