@Eric-Arellano 2021-09-09
Thanks so much @williamscs! FYI I merged with the title "Fix backports of std lib modules like `dataclasses` not working with dependency inference" - the PR title was already excellent, but changed so that we can categorize this as a "bug fix" in the changelog.

I'll cherry-pick to 2.7 now